Refactored POSTGRES environment variables
The jdbc host address was set to POSTGRES_PORT_5432_TCP_ADDR. Updated it to POSTGRES_HOST and set the default to be the link name in the README (postgres).

The jdbc port was set to use env.POSTGRES_PORT_5432_TCP_PORT when the README specified POSTGRES_PORT as an environment variable.

These issues prevent keycloak from deploying.