Move sed and fix typo in server/README.md
Move the `sed` command below the definition of `JBOSS_HOME`.  It appears
standalone.xml doesn't exist at the point when the `sed` was run.
Additionally the environment variable `JBOSS_HOME` was not defined.

Also added the equals sign to the `server/README.md`. Without an equals
sign docker assumes true is the container name which causes it to fail.