Fixing truststore import message
I was using the `X509_CA_BUNDLE` feature when I noticed in the logs that it `"Failed to import certificates from system's Java CA certificate bundle into Keycloak truststore!"`.

I pulled the container down and built the truststore manually like the script does, and it succeeded. I checked the return code and it was 0, so I believe this is backwards.