KEYCLOAK-12387 Add jboss group
With KEYCLOAK-10059 the base image has been changed to ubi8-minimal.
Since ubi8-minimal does not ship with a user named 'jboss', the user
will be created in the startup scripts. However, the previously used
image jboss/base-jdk also ships with a group named 'jboss', the
ubi8-minimal image does not. With KEYCLOAK-10059 this group
does not get created

For backwards compatibility a group named 'jboss' has been added and
the user 'jboss' has been associated as a member of that group.